Handles DateTimes without relying on estimatedWait
scheduledTime is converted into a Date object via a Calendar. This allows
us to check when times are also supplied for the next day (ie if a request
is made at 23:59, there will be times for 00:00 onwards!). This will be
more important when it comes to data caching, as it will facilitate
checking for cache hits. This also removes the estimatedWait hack.